Created buildable turrets
