--- Assets/BuildingConstructionPreview.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1c80a645275000c4e854588a19646b9b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
