+++ Assets/Prefabs/Robot Empire/Buildings/Base Defense/Base_Defenses_Building_Template.prefab
@@ -295,7 +295,7 @@ MonoBehaviour:
     m_BestFit: 0
     m_MinSize: 0
     m_MaxSize: 89
-    m_Alignment: 0
+    m_Alignment: 1
     m_AlignByGeometry: 0
     m_RichText: 1
     m_HorizontalOverflow: 1
