+++ Assets/Prefabs/Robot Empire/Buildings/Troop Production/Troop_Production_Building_Template.prefab
@@ -446,7 +446,7 @@ MonoBehaviour:
     m_BestFit: 0
     m_MinSize: 0
     m_MaxSize: 89
-    m_Alignment: 0
+    m_Alignment: 1
     m_AlignByGeometry: 0
     m_RichText: 1
     m_HorizontalOverflow: 1
