+++ Assets/Prefabs/Robot Empire/Buildings/Turret.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cb3a922b9c6f62949bcdd0993aa62f97
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
