+++ Assets/Prefabs/Robot Empire/Buildings/Turret/Turret_Building_Model.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 654d75b360ec0da40b04b0ba29521c28
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
