+++ Assets/Prefabs/Robot Empire/Buildings/Turret/Turret_Building_Preview.prefab
@@ -1,0 +1,125 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!1 &7634239093149715558
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 5723388860028118715}
+  - component: {fileID: 7927317092807639052}
+  m_Layer: 0
+  m_Name: Turret_Building_Preview
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &5723388860028118715
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 7634239093149715558}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 1271.06, y: 17.28, z: -646.77}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_Children:
+  - {fileID: 8226002571731311962}
+  m_Father: {fileID: 0}
+  m_RootOrder: 0
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &7927317092807639052
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 7634239093149715558}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 1c80a645275000c4e854588a19646b9b, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  buildingPrefab: {fileID: 8421837390822842998, guid: 9b45ed4e007d8704d92ccb01917306a8,
+    type: 3}
+  allowed: {fileID: 2100000, guid: efcd7f6f8eb876f4e85235b09cdf7b23, type: 2}
+  denied: {fileID: 2100000, guid: a2a46c986fd34c64ba50d56961df6e15, type: 2}
+--- !u!1001 &8388416440678282555
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 5723388860028118715}
+    m_Modifications:
+    - target: {fileID: 450647546899097697, guid: 654d75b360ec0da40b04b0ba29521c28,
+        type: 3}
+      propertyPath: m_RootOrder
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 450647546899097697, guid: 654d75b360ec0da40b04b0ba29521c28,
+        type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 450647546899097697, guid: 654d75b360ec0da40b04b0ba29521c28,
+        type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 7.31
+      objectReference: {fileID: 0}
+    - target: {fileID: 450647546899097697, guid: 654d75b360ec0da40b04b0ba29521c28,
+        type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 450647546899097697, guid: 654d75b360ec0da40b04b0ba29521c28,
+        type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 450647546899097697, guid: 654d75b360ec0da40b04b0ba29521c28,
+        type: 3}
+      propertyPath: m_LocalRotation.x
+      value: -0
+      objectReference: {fileID: 0}
+    - target: {fileID: 450647546899097697, guid: 654d75b360ec0da40b04b0ba29521c28,
+        type: 3}
+      propertyPath: m_LocalRotation.y
+      value: -0
+      objectReference: {fileID: 0}
+    - target: {fileID: 450647546899097697, guid: 654d75b360ec0da40b04b0ba29521c28,
+        type: 3}
+      propertyPath: m_LocalRotation.z
+      value: -0
+      objectReference: {fileID: 0}
+    - target: {fileID: 450647546899097697, guid: 654d75b360ec0da40b04b0ba29521c28,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 450647546899097697, guid: 654d75b360ec0da40b04b0ba29521c28,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 450647546899097697, guid: 654d75b360ec0da40b04b0ba29521c28,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3480482383476919482, guid: 654d75b360ec0da40b04b0ba29521c28,
+        type: 3}
+      propertyPath: m_Name
+      value: Turret_Building_Model
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: 654d75b360ec0da40b04b0ba29521c28, type: 3}
+--- !u!4 &8226002571731311962 stripped
+Transform:
+  m_CorrespondingSourceObject: {fileID: 450647546899097697, guid: 654d75b360ec0da40b04b0ba29521c28,
+    type: 3}
+  m_PrefabInstance: {fileID: 8388416440678282555}
+  m_PrefabAsset: {fileID: 0}
