+++ Assets/Prefabs/Robot Empire/Buildings/Turret/Turret_Building_Preview.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 968cabca3cb02d94788797c38cf9b0fd
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
