+++ Assets/Prefabs/Robot Empire/Buildings/Turret/Turret_Building_Template.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 9b45ed4e007d8704d92ccb01917306a8
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
