+++ Assets/Scripts/Construction/BuildingConstruction.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bc34b34ae2d843b4e936aa40b11f9160
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
