+++ Assets/Scripts/Construction/Building_Name.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 43d99cc5a002ef649afae5746ecc3fb6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
